adapted the joystick gui to the new resolution of the ball handle message

+++ msl_joystick/include/msl_joystick/Joystick.h
@@ -75,8 +75,8 @@ namespace msl_joystick
 		vector<bool> keyPressed;
 
 		// for filling a joystick message
-		short ballHandleLeftMotor;
-		short ballHandleRightMotor;
+		int ballHandleLeftMotor;
+		int ballHandleRightMotor;
 		short kickPower;
 		int robotId;
 		bool useBallHandle;
@@ -85,8 +85,8 @@ namespace msl_joystick
 		double rotation;
 
 		// min max values from config
-		short ballHandleMin;
-		short ballHandleMax;
+		int ballHandleMin;
+		int ballHandleMax;
 		short kickPowerMin;
 		short kickPowerMax;
 		double translationMin;
